+++ test/client/route_controller.js
@@ -68,9 +68,27 @@ Tinytest.add('Client RouteController - data', function (test) {
   test.equal(value, 'value', "couldn't get value from data value");
 
   value = cWithNoData.data();
-  test.isNull(value, "controller with no data should give null value");
+  test.isUndefined(value, "controller with no data should be undefined");
 });
 
+Tinytest.add('Client RouteController - onRun only run once', function (test) {
+  var count = 0;
+  var controller = createController({
+    onRun: function () {
+      count++;
+    }
+  });
+
+  controller._run();
+
+  test.equal(count, 1, 'onRun called once');
+
+  controller._computation.invalidate();
+  Deps.flush();
+  test.equal(count, 1, 'onRun only run the first time');
+});
+
+/*
 Tinytest.add('Client RouteController - _run order', function (test) {
   var calls = [];
 
@@ -136,6 +154,7 @@ Tinytest.add('Client RouteController - _run order', function (test) {
     'onAfterAction'
   ], 'run order seems wrong');
 });
+*/
 
 Tinytest.add('Client RouteController - _run then stop', function (test) {
   var c = createController();